+++ embrace/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 4.8.0
 
 * Version bump
+* Added a screen-load span to `EmbraceNavigationObserver`, timed from the last user touch to the first frame of the new route
 
 # 4.7.0